requirements.txt
--extra-index-url https://storage.openvinotoolkit.org/simple/wheels/pre-release
--extra-index-url https://storage.openvinotoolkit.org/simple/wheels/nightly
openvino_genai~=2025.1.0.0.dev
librosa==0.10.2.post1  # For Whisper
pillow==11.1.0  # Image processing for VLMs
--extra-index-url https://download.pytorch.org/whl/cpu
--extra-index-url https://storage.openvinotoolkit.org/simple/wheels/pre-release
--extra-index-url https://storage.openvinotoolkit.org/simple/wheels/nightly
openvino-tokenizers~=2025.1.0.0.dev
optimum-intel @ git+https://github.com/huggingface/optimum-intel.git@main
numpy<2.0.0; sys_platform == 'darwin'
einops==0.8.1  # For Qwen
transformers_stream_generator==0.0.5  # For Qwen
diffusers==0.32.2 # For image generation pipelines
timm==1.0.15  # For exporting InternVL2
torchvision  # For visual language models
transformers>=4.43 # For Whisper
hf_transfer # for faster models download, should used with env var HF_HUB_ENABLE_HF_TRANSFER=1
gradio==5.21.0